+++ src/components/search-line.js
@@ -1,24 +1,6 @@
-import {createElement} from './utils';
+import {AbstractComponent} from './abstract-component';
 
-export class SearchLine {
-  constructor() {
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    if (this._element) {
-      this._element = null;
-    }
-    return this._element;
-  }
-
+export class SearchLine extends AbstractComponent {
   getTemplate() {
     return `<form class="header__search search">
         <input type="text" name="search" class="search__field" placeholder="Search movies">
