+++ src/components/sorting-line.js
@@ -1,24 +1,6 @@
-import {createElement} from './utils';
+import {AbstractComponent} from './abstract-component';
 
-export class SortingLine {
-  constructor() {
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    if (this._element) {
-      this._element = null;
-    }
-    return this._element;
-  }
-
+export class SortingLine extends AbstractComponent{
   getTemplate() {
     return `
       <ul class="sort">
